requirements.txt
# python.exe -m pip install --upgrade pip
google-generativeai
python-dotenv
"--upgrade grpcio"
streamlit